+++ packages/client/src/App.tsx
@@ -1,13 +1,13 @@
-import { useEffect } from 'react';
+// import { useEffect } from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import PageNavigation from './components/page-navigation/page-navigation';
 import { Main } from './pages/main';
-import { Login } from './pages/login/login';
-import { SignUp } from './pages/sign-up/sign-up';
-import NewGame from './pages/new-game';
-import Profile from './pages/profile';
-import Rules from './pages/rules';
-import Leaders from './pages/leaders';
+import { Login } from './pages/login';
+import { SignUp } from './pages/sign-up';
+import { NewGame } from './pages/new-game';
+import { Profile } from './pages/profile';
+import { Rules } from './pages/rules';
+import { Leaders } from './pages/leaders';
 
 import './scss/style.scss';
