+++ packages/client/src/pages/forum.tsx
@@ -1,11 +1,9 @@
 import React from 'react';
 
-const Forum: React.FC = (): JSX.Element => {
+export const Forum: React.FC = (): JSX.Element => {
   return (
     <main>
       <h1>Форум</h1>
     </main>
   )
 }
-
-export default Forum
