+++ packages/client/src/pages/leaders.tsx
@@ -1,11 +1,9 @@
 import React from 'react';
 
-const Leaders: React.FC = (): JSX.Element => {
+export const Leaders: React.FC = (): JSX.Element => {
   return (
     <main>
       <h1>Лидеры</h1>
     </main>
   )
 }
-
-export default Leaders
