+++ packages/client/src/pages/new-game.tsx
@@ -1,11 +1,9 @@
 import React from 'react';
 
-const NewGame: React.FC = (): JSX.Element => {
+export const NewGame: React.FC = (): JSX.Element => {
   return (
     <main>
       <h1>Новая игра</h1>
     </main>
   )
 }
-
-export default NewGame
