+++ packages/client/src/pages/profile.tsx
@@ -1,11 +1,9 @@
 import React from 'react';
 
-const Profile: React.FC = (): JSX.Element => {
+export const Profile: React.FC = (): JSX.Element => {
   return (
     <main>
       <h1>Профиль</h1>
     </main>
   )
-}
-
-export default Profile
+};
