+++ packages/client/src/pages/rules.tsx
@@ -1,11 +1,9 @@
 import React from 'react';
 
-const Rules: React.FC = (): JSX.Element => {
+export const Rules: React.FC = (): JSX.Element => {
   return (
     <main>
       <h1>Правила</h1>
     </main>
   )
 }
-
-export default Rules
